+++ src/IECoreRenderMan/Globals.cpp
@@ -564,7 +564,7 @@ void Globals::pause()
 
 void Globals::updateRenderView()
 {
-	// Find camera.
+	// Find camera and update options from it.
 
 	Session::CameraInfo camera = m_session->cameraInfo( m_cameraOption );
 	if( camera.id == riley::CameraId::InvalidId() )
@@ -587,6 +587,20 @@ void Globals::updateRenderView()
 			);
 		}
 		camera.id = m_defaultCamera;
+	}
+
+	if( std::any_of( m_outputs.begin(), m_outputs.end(), [] ( const auto &v ) { return v.second->getType() == "quicklyNoiseless"; } ) )
+	{
+		// The `quicklyNoiseless` driver doesn't handle interactive edits to the
+		// crop window - it variously crashes, offsets the image, or fails to clear
+		// the area outside the data window. So when the crop changes we delete the
+		// render view and create new drivers from scratch.
+		const float *oldCropWindow = m_options.GetFloatArray( Loader::strings().k_Ri_CropWindow, 4 );
+		const float *newCropWindow = camera.options.GetFloatArray( Loader::strings().k_Ri_CropWindow, 4 );
+		if( oldCropWindow && newCropWindow && !std::equal( oldCropWindow, oldCropWindow + 4, newCropWindow ) )
+		{
+			deleteRenderView();
+		}
 	}
 
 	m_options.Update( camera.options );